Add static cast to float for M_PI

+++ src/hand-tracking/src/VisualProprioception.cpp
@@ -196,7 +196,7 @@ VisualProprioception::VisualProprioception(const bool use_thumb,
         si_cad_ = new SICAD(cad_obj_,
                             cam_width_, cam_height_,cam_fx_, cam_fy_, cam_cx_, cam_cy_,
                             num_images,
-                            {1.0, 0.0, 0.0, M_PI},
+                            {1.0, 0.0, 0.0, static_cast<float>(M_PI)},
                             shader_path,
                             false);
     }
